refactor: フロントエンドでも依存するのであれば抽象モデルであるべき
