+++ server/models/searchQuery.ts
